% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BcAbundance}
\alias{BcAbundance}
\title{Count barcode abundances per individual and according conditions}
\usage{
BcAbundance(longMatrix, metadata, indivVar, listVar, colorVar = "")
}
\arguments{
\item{longMatrix, }{dataframe in long format}

\item{metadata, }{the metadata}

\item{indivVar, }{variable corresponding to indivuals variable}

\item{listVar, }{a list of variables that correspond to metadata column names}

\item{colorVar, }{color variable name}
}
\value{
a long matrix with barcode abundances in each set of variable choose.
}
\description{
Count barcode abundances per individual and according conditions
}
