% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CalculDiversity}
\alias{CalculDiversity}
\title{Calucul barcodes diversity}
\usage{
CalculDiversity(
  matrix,
  metadata,
  indivVar,
  indivVal,
  listVar,
  listVal,
  colorVar = "",
  diversity = "Number of barcodes"
)
}
\arguments{
\item{matrix, }{a matrix}

\item{metadata, }{the metadata}

\item{indivVar, }{individuals variable name}

\item{indivVal, }{individuals value names}

\item{listVar, }{a list of varaibles}

\item{listVal, }{a list of values}

\item{colorVar}{, the color variable}

\item{diversity, }{diversity index, default is "Number of barcodes". Others are "Simpson Index" or "Shannon Index"}
}
\value{
a matrix with diversity values
}
\description{
Calucul barcodes diversity
}
