% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CheckValueLengths}
\alias{CheckValueLengths}
\title{Check that each variable has at least one value}
\usage{
CheckValueLengths(listVar, listVal, metadata)
}
\arguments{
\item{listVar, }{list of selected variable(s)}

\item{listVal, }{list of selected value(s)}

\item{metadata, }{metadata tables with variable and value names}
}
\value{
an integer
}
\description{
Check that each variable has at least one value
}
