% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ColToRowNames}
\alias{ColToRowNames}
\title{First column as rownames}
\usage{
ColToRowNames(matrix)
}
\arguments{
\item{matrix, }{a matrix with as first column, row names}
}
\value{
a matrix with row names previously in th efirst column
}
\description{
First column as rownames
}
