% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DeleteLevels}
\alias{DeleteLevels}
\title{Delete levels in a list}
\usage{
DeleteLevels(list)
}
\arguments{
\item{list, }{a list}
}
\value{
a list without levels.
}
\description{
Delete levels in a list
}
