% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ListToDf}
\alias{ListToDf}
\title{Convert a list of several lists to a dataframe}
\usage{
ListToDf(list)
}
\arguments{
\item{list, }{a list of lists}
}
\value{
a data-frame
}
\description{
Convert a list of several lists to a dataframe
}
