% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LongSubMatrix}
\alias{LongSubMatrix}
\title{Select values and individuals}
\usage{
LongSubMatrix(longMatrix, indivVar, indivVal, listVar, listVal, metadata)
}
\arguments{
\item{longMatrix, }{a long matrix with sample names splited into columns}

\item{indivVar, }{individuals variable names}

\item{indivVal, }{individuals value names in same order than variables}

\item{listVar, }{a list of variables that correspond to metadata column names}

\item{listVal, }{a list of values that correspond to metadata values}

\item{metadata, }{a dataframe containing variables as columns names and their corresponding values}
}
\value{
a long df with samples containing values gave in listVal
}
\description{
Select values and individuals
}
