% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LongToWideSubMatrix}
\alias{LongToWideSubMatrix}
\title{Select samples according user selection on variables and individuals}
\usage{
LongToWideSubMatrix(longMatrix, metadata, indivVar, indivVal, listVar, listVal)
}
\arguments{
\item{longMatrix, }{long format matrix splitted sample names in columns}

\item{metadata, }{a dataframe containing variables as columns names and their corresponding values}

\item{indivVar, }{individuals variable name}

\item{indivVal, }{individuals value name}

\item{listVar, }{a list of variables that correspond to metadata column names}

\item{listVal, }{a list of values that correspond to metadata values}
}
\value{
a long dataframe matrix with samples containing values gave in listVal
}
\description{
Select samples according user selection on variables and individuals
}
