% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MakeBarcodeFrequenciesMatrix}
\alias{MakeBarcodeFrequenciesMatrix}
\title{Plot cumulative diagram}
\usage{
MakeBarcodeFrequenciesMatrix(
  matrix,
  metadata,
  indivVar,
  indivVal,
  listVar,
  listVal
)
}
\arguments{
\item{matrix, }{wide format barcode count matrix}

\item{metadata, }{a dataframe containing variables as columns names and their corresponding values}

\item{indivVar, }{individuals variable name}

\item{indivVal, }{individuals value name}

\item{listVar, }{a list of variables that correspond to metadata column names}

\item{listVal, }{a list of values that correspond to metadata values}
}
\value{
a cumulative plot
}
\description{
Plot cumulative diagram
}
