% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MakeCategoryMatrices}
\alias{MakeCategoryMatrices}
\title{Make catergory matrices}
\usage{
MakeCategoryMatrices(
  matrixWide,
  metadata,
  indivVar,
  indivVal,
  cellTypeVar,
  cellTypeVal,
  threshold = 10,
  condition = "no",
  conditionVal = ""
)
}
\arguments{
\item{metadata, }{the metadata corresponding to the matrix}

\item{indivVar, }{name of variable defining individuals}

\item{indivVal, }{list of selected individuals}

\item{cellTypeVar, }{list of sected variables}

\item{cellTypeVal, }{list of sected values}

\item{threshold, }{lineage bias minimum threshold (\%), default is 10}

\item{condition, }{is there a condtion, default is "no"}

\item{conditionVal, }{condition value (only one accepted), default is ""}

\item{matrix, }{barcode count matrix}
}
\value{
a list of two matrices, the category for PlotCategories function and barcode count matrix for PlotCategoryCount function
}
\description{
Make catergory matrices
}
