% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MakeCumulativeDiagramMatrix}
\alias{MakeCumulativeDiagramMatrix}
\title{Plot and create matrix of a cumulative diagram}
\usage{
MakeCumulativeDiagramMatrix(
  matrix,
  metadata,
  indivVar,
  indivVal,
  var,
  listVal,
  colorVar = "",
  textSize = 15,
  xProp = "no"
)
}
\arguments{
\item{matrix, }{a long matrix}

\item{metadata, }{the metadta}

\item{indivVar, }{individuals variable name}

\item{indivVal, }{individuals values name}

\item{var, }{a variable}

\item{listVal, }{list of variable values wanted}

\item{colorVar, }{variable to split and color samples by}

\item{textSize, }{size of the plot text, deault is 15}

\item{xProp, }{should the x axis be in proportions, default is "no" Other is "yes"}
}
\value{
a matrix to generate the cumulative plot
}
\description{
Plot and create matrix of a cumulative diagram
}
