% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MakeDotPlotMatrix}
\alias{MakeDotPlotMatrix}
\title{Make matrix for PlotDotPlot function}
\usage{
MakeDotPlotMatrix(
  matrixWide,
  metadata,
  indivVar,
  indivVal,
  xVar,
  xVal,
  yVar,
  yVal,
  colorVar = "",
  transformation = "arcsin"
)
}
\arguments{
\item{matrixWide, }{a barcode count matrix}

\item{metadata, }{the metadata}

\item{indivVar, }{individual variable name}

\item{xVar, }{x axis variable}

\item{xVal, }{x axis values}

\item{yVar, }{y axis variable}

\item{yVal, }{y axis values}

\item{colorVar, }{color variable as column in the matrix}
}
\value{
a matrix
}
\description{
Make matrix for PlotDotPlot function
}
