% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MakeDuplicatesMatrix}
\alias{MakeDuplicatesMatrix}
\title{Select values to create duplicates matrix}
\usage{
MakeDuplicatesMatrix(matrix, listVar, listVal, metadata)
}
\arguments{
\item{matrix, }{a long matrix made with ReformatQCmatrix function}

\item{listVar, }{a list of variables that correspond to metadata column names}

\item{listVal, }{a list of values that correspond to metadata values}

\item{metadata, }{a dataframe containing variables as columns names and their corresponding values}
}
\value{
a long df with samples containing values gave in listVal
}
\description{
Select values to create duplicates matrix
}
