% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MakeHeatmapMatrix}
\alias{MakeHeatmapMatrix}
\title{Make input matrix for heatmap}
\usage{
MakeHeatmapMatrix(
  matrix,
  metadata,
  indivVar,
  indivVal = "",
  listVar,
  listVal,
  poolIndiv = TRUE
)
}
\arguments{
\item{matrix, }{a barcode count matrix}

\item{metadata, }{the metadata}

\item{indivVar, }{variable name defining individuals}

\item{indivVal, }{list of values, default is "" because of poolIndiv=TRUE}

\item{listVar, }{list of variables}

\item{listVal, }{list of values}

\item{poolIndiv, }{should individuals be pooled, default is TRUE}
}
\value{
a matrix to as input to PlotHeatmap function
}
\description{
Make input matrix for heatmap
}
