% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MakeMetadata}
\alias{MakeMetadata}
\title{Generate your metadata}
\usage{
MakeMetadata(matrix, separataor = "_")
}
\arguments{
\item{matrix, }{a barcode count matrix (samples by columns and barcodes by rows)}

\item{separataor, }{fileds separator used to split sample names, default is "_"}
}
\value{
a data-frame
}
\description{
Generate your metadata
}
