% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MakePieChartMatrix}
\alias{MakePieChartMatrix}
\title{Make input matrix for PlotPieChart function}
\usage{
MakePieChartMatrix(matrix, indivVar, colorVar)
}
\arguments{
\item{matrix, }{matrix made with MakeDotPlotMatrix function}

\item{indivVar, }{individual variable name}

\item{colorVar, }{color variable as column in the matrix}

\item{textSize, }{default is 8}
}
\value{
a matrix for PlotPieChart
}
\description{
Make input matrix for PlotPieChart function
}
