% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MakeRepeatUseMatrix}
\alias{MakeRepeatUseMatrix}
\title{Make RepeatUse Matrix}
\usage{
MakeRepeatUseMatrix(matrix, indivVar, indivVal)
}
\arguments{
\item{matrix, }{a wide dataframe created with ReformatQCmatrix function}

\item{indivVar, }{indiviudal variable name}

\item{indivVal, }{indiviudal value names}
}
\value{
matrix
}
\description{
Make RepeatUse Matrix
}
