% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MakeTernaryMatrix}
\alias{MakeTernaryMatrix}
\title{Create matrix input for the PlotTernaryPlot function}
\usage{
MakeTernaryMatrix(
  matrixWide,
  metadata,
  indivVar,
  indivVal,
  topVar,
  topVal,
  rightVar,
  rightVal,
  leftVar,
  leftVal,
  addColor = "no"
)
}
\arguments{
\item{matrixWide, }{barcode count matrix}

\item{metadata, }{the metadata of the matrix}

\item{indivVar, }{individuals variable name}

\item{indivVal, }{individuals value names}

\item{topVar, }{top ternary variable}

\item{topVal, }{top ternary values}

\item{rightVar, }{right ternary variable}

\item{rightVal, }{right ternary values}

\item{leftVar, }{left ternary variable}

\item{leftVal, }{left ternary values}

\item{addColor, }{should the dots be colored by individuals, default is "no". Other is "yes"}
}
\value{
a ternary plot
}
\description{
Create matrix input for the PlotTernaryPlot function
}
