% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NormByBcAbundance}
\alias{NormByBcAbundance}
\title{Normalize barcode abundances by their total abundance (by row)}
\usage{
NormByBcAbundance(matrix, nbVal)
}
\arguments{
\item{matrix, }{a dataframe}

\item{nbVal, }{number of values}
}
\value{
a matrix with normalized barcode abundances by their total abundance (in \%)
}
\description{
Normalize barcode abundances by their total abundance (by row)
}
