% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotBarcodeFrequencies}
\alias{PlotBarcodeFrequencies}
\title{Plot a frequency diagram}
\usage{
PlotBarcodeFrequencies(
  subLgMatrix,
  colorVar = "",
  y = "density",
  nbins = 5,
  log = "no",
  textSize = 15
)
}
\arguments{
\item{subLgMatrix, }{a long matrix}

\item{colorVar, }{variable to split and color samples by}

\item{y, }{should the frequencies by ploted in a density curve or an histogram, degault is "density". Other is "histogram"}

\item{nbins, }{number of bins, deauflt is 5}

\item{log, }{shoudl the x axis be in log scale, deault is "no". Other is "yes"}

\item{textSize, }{size of plot text, default is 15}
}
\value{
a frequency density curve or frequency histogram plot
}
\description{
Plot a frequency diagram
}
