% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotBias}
\alias{PlotBias}
\title{Plot a bias plot}
\usage{
PlotBias(biasMatx, conditionVal = "", textSize = 15)
}
\arguments{
\item{biasMatx, }{the matrix made with MakeBiasMatrix()}

\item{conditionVal}{, condition value (only one accepted), the same as in MakeBiasMatrix(). Default is "".}

\item{textSize}{, the sier of the text, default is 15.}
}
\value{
a bargraphe
}
\description{
Plot a bias plot
}
