% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotBiasPerCat}
\alias{PlotBiasPerCat}
\title{Plot a bias plot}
\usage{
PlotBiasPerCat(biasMatx, conditionVal = "", textSize = 15)
}
\arguments{
\item{biasMatx, }{the matrix made with MakeBiasPerCatMatrix()}

\item{conditionVal}{, condition value (only one accepted), the same as in MakeBiasPerCatMatrix(). Default is "".}

\item{textSize}{, the sier of the text, default is 15.}
}
\value{
a bargraphe
}
\description{
Plot a bias plot
}
