% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotBiasPerType}
\alias{PlotBiasPerType}
\title{Plot a bias plot}
\usage{
PlotBiasPerType(biasMatx, y, conditionVal = "", textSize = 15)
}
\arguments{
\item{biasMatx, }{the matrix made with MakeBiasPerTypeMatrix()}

\item{y}{, the cell type that you want to plot in y axis. It has to be one of the cellTypeVal used in MakeBiasPerTypeMatrix()}

\item{conditionVal}{, condition value (only one accepted), the same as in MakeBiasPerTypeMatrix(). Default is "".}

\item{textSize}{, the sier of the text, default is 15.}
}
\value{
a bargraphe
}
\description{
Plot a bias plot
}
