% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotCategories}
\alias{PlotCategories}
\title{Plot a category bias plot}
\usage{
PlotCategories(
  catMatx,
  threshold = 10,
  conditionVal = "",
  textSize = 15,
  legendPos = "right"
)
}
\arguments{
\item{catMatx}{, the first matrix of the list created with MakeCategoryMatrices function}

\item{threshold}{, the lineage bias threshold, default is 10. Same used for MakeCategoryMatrices function.}

\item{conditionVal}{, value of the condition, default is "". Same used for MakeCategoryMatrices function.}

\item{textSize}{, the sier of the text, default is 15.}

\item{legendPos}{, position of the legend arround the plot, default is right. Other options are "bottom" or "top" or "left".}
}
\value{
a bargraphe
}
\description{
Plot a category bias plot
}
