% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotCategoryCounts}
\alias{PlotCategoryCounts}
\title{Plot a category counts plot}
\usage{
PlotCategoryCounts(
  catCountMatx,
  threshold = 10,
  conditionVal = "",
  textSize = 15
)
}
\arguments{
\item{catCountMatx, }{the second matrix of the list created with MakeCategoryMatrices function}

\item{threshold}{, the lineage bias threshold, default is 10}

\item{conditionVal, }{condition value, default is ""}

\item{textSize}{, the sier of the text, default is 15}
}
\value{
a bargraphe
}
\description{
Plot a category counts plot
}
