% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotCorrelogram}
\alias{PlotCorrelogram}
\title{Plot a correlogram}
\usage{
PlotCorrelogram(matrix, correlation = "pearson")
}
\arguments{
\item{matrix, }{a matrix with barcode as rownames and values as columns}

\item{correlation, }{correlation method between "pearson and "spearman", default is "pearson"}
}
\value{
a correlogram
}
\description{
Plot a correlogram
}
