% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotCumulativeDiagram}
\alias{PlotCumulativeDiagram}
\title{Plot cumulative diagram}
\usage{
PlotCumulativeDiagram(
  matrix,
  indivVar,
  colorVar = "",
  xProp = "no",
  textSize = 15
)
}
\arguments{
\item{matrix, }{matrix obtained with MakeCumulativeDiagramMatrix function}

\item{indivVar, }{individuals variable name}

\item{colorVar, }{variable to split and color samples by}

\item{xProp, }{should the x axis be in proportions, default is "no". Other is "yes"}

\item{textSize, }{size of the plot text, deault is 15}
}
\value{
a cumulative plot
}
\description{
Plot cumulative diagram
}
