% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotDiversity}
\alias{PlotDiversity}
\title{Plot a boxplot with diversity measures}
\usage{
PlotDiversity(
  matrix,
  diversity,
  listVar,
  indivVar,
  colorVar = "",
  dots = "no",
  labels = "no",
  textSize = 15
)
}
\arguments{
\item{matrix, }{a long matrix containing diversity measures make with CalculDiversity function}

\item{diversity, }{diversity index, default is "Number of barcodes". Others are "Simpson Index" or "Shannon Index"}

\item{listVar}{, a list of varaibles. Same used in CalculDiversity function.}

\item{indivVar, }{individuals variable name. Same used in CalculDiversity function.}

\item{colorVar}{, the color variable. Same used in CalculDiversity function.}

\item{dots, }{reveal dots, default is "no". Other is "yes"}

\item{labels, }{show dots labels, default is "no". Other is "yes".}

\item{textSize, }{size of the text, default is 15}
}
\value{
a boxplot
}
\description{
Plot a boxplot with diversity measures
}
