% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotDotplot}
\alias{PlotDotplot}
\title{Plot a dotplot}
\usage{
PlotDotplot(
  longMatrix,
  indivVar,
  colorVar = "",
  transformation = "arcsin",
  textSize = 15
)
}
\arguments{
\item{longMatrix, }{a matrix made with MakeDotPlotMatrix function}

\item{indivVar, }{individual variable name}

\item{colorVar, }{variable defining your individuals only}

\item{transformation, }{transformation to use, default is "arcsin". Other is "log10"}

\item{textSize, }{size of the text of the plot, deault is 15}
}
\value{
a dotplot of barcode abundances in two conditions
}
\description{
Plot a dotplot
}
