% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotDuplicates}
\alias{PlotDuplicates}
\title{Plot Duplicates}
\usage{
PlotDuplicates(
  matrix,
  dupVal,
  transformation = "arcsin",
  textSize = 15,
  correlation = "spearman"
)
}
\arguments{
\item{matrix, }{a made by 1)ReformatQCmatrix --> 2) MakeDuplicatesMatrix functions}

\item{dupVal, }{duplicate value names as a list}

\item{transformation, }{transformation name, default is "arcsin". Other "log10(x+1)" or "none". It has to be the same as used for ReformatQCmatrix function}

\item{textSize, }{size of text, default is 15}

\item{correlation, }{the correlation used in ReformatQCmatrix() function. "spearman" (default) or "pearson"}
}
\value{
a Duplicates matrix
}
\description{
Plot Duplicates
}
