% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotHeatmap}
\alias{PlotHeatmap}
\title{Plot a heatmap}
\usage{
PlotHeatmap(
  wideMatrix,
  distance = "euclidean",
  clustering = "complete",
  showDendro = "no",
  showBarcodes = "no",
  nClusters = 0,
  columnTextSize = 1
)
}
\arguments{
\item{wideMatrix, }{a dataframe treated with LongToWideSubMatrix() or LongToWideSubMatrix_pooledIndiv() functions}

\item{distance, }{heatmap distance metrics}

\item{clustering, }{heatmap clustering method}

\item{showDendro, }{"yes" or "no". If yes, column dendrogram is shown}

\item{showBarcodes, }{"yes" or "no". If yes, row barcode names are shown}

\item{nClusters, }{number of clusters to be shown, minimum is 0}

\item{columnTextSize, }{column text size}
}
\value{
a heatmap of samples in {wideMatrix}
}
\description{
Plot a heatmap
}
