% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotPieChart}
\alias{PlotPieChart}
\title{Plot a piechart of the number of shared or specific barcodes}
\usage{
PlotPieChart(matrix, textSize = 8)
}
\arguments{
\item{matrix, }{matrix made with MakePieChartMatrix function}

\item{textSize, }{default is 8}
}
\value{
a piechart
}
\description{
Plot a piechart of the number of shared or specific barcodes
}
