% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotTernaryPlot}
\alias{PlotTernaryPlot}
\title{Plot a ternary plot}
\usage{
PlotTernaryPlot(matrix, indivVar, addColor = "no", textSize = 15)
}
\arguments{
\item{matrix, }{a dataframe created by MakeTernaryMatrix function}

\item{indivVar, }{individuals variable name, same as used in MakeTernaryMatrix function}

\item{addColor, }{should the dots be colored by individuals, default is "no". Other is "yes"}

\item{textSize, }{text size, default is 15}
}
\value{
a ternary plot
}
\description{
Plot a ternary plot
}
