% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ReformatQCmatrix}
\alias{ReformatQCmatrix}
\title{Reformat QC input table}
\usage{
ReformatQCmatrix(
  matrix,
  metadata,
  dupVar,
  dupVal,
  sampleNameFieldsep = "_",
  transformation = "arcsin",
  correlation = "spearman"
)
}
\arguments{
\item{matrix, }{a wide dataframe}

\item{metadata, }{the metadata}

\item{dupVar, }{duplicate variable name}

\item{dupVal, }{duplicate values}

\item{sampleNameFieldsep, }{sample name variable separatorn default "_"}

\item{transformation, }{transformation to use, default is "arcsin". Other "log10(x+1)" or "none"}

\item{correlation, }{correlation, "spearman" (default) or "pearson"}
}
\value{
a long format QC matrix
}
\description{
Reformat QC input table
}
