% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SumVars}
\alias{SumVars}
\title{Create new columns with stitched value names and summed listVal}
\usage{
SumVars(matx, listVal)
}
\arguments{
\item{matx, }{a long matrix with barcodes + individuals + value to stich together}

\item{listVal, }{list of values to stitch}
}
\value{
a matrix with stiched value as column names and summed counts
}
\description{
Create new columns with stitched value names and summed listVal
}
