% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{WideToLong}
\alias{WideToLong}
\title{Transform a wide to long matrix}
\usage{
WideToLong(wideMatrix, metadata)
}
\arguments{
\item{wideMatrix, }{a wide format matrix with samples by columns and barcode by rows}

\item{metadata, }{the metadata of the matrix}
}
\value{
a long format matrix with sample names decomposed with metadata column names
}
\description{
Transform a wide to long matrix
}
