% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{PlotRepeatUse}
\alias{PlotRepeatUse}
\title{Plot a dotplot of barcode abundances between individuals}
\usage{
PlotRepeatUse(matrix, indivVar, textSize = 15)
}
\arguments{
\item{matrix, }{a wide dataframe with individuals by columns}

\item{indivVar, }{indiviudal variable name}

\item{textSize, }{text size, default is 15}

\item{indivCombinations, }{a dataframe of all 2D individual combinations}
}
\value{
one or several dotplots of barcode abundances in two individuals
}
\description{
Plot a dotplot of barcode abundances between individuals
}
